Moved the capability logic into a capability util file.
