+++ how-to/workspace-platform-starter/client/src/framework/snap.ts
@@ -7,15 +7,14 @@ import { MANIFEST_TYPES } from "./manifest-types";
 import * as platformSplashProvider from "./platform/platform-splash";
 import type { SnapProviderOptions } from "./shapes";
 import { formatError, isEmpty } from "./utils";
+import { getCanDownloadAppAssets, getCanLaunchExternalProcess } from "./utils-capability";
 
 const logger = createLogger("Snap");
 const NATIVE_APP_PREFIX = "app:";
 
 let server: Snap.SnapServer | undefined;
 let isSnapEnabled: boolean = false;
-let canLaunchExternalProcess: boolean | undefined;
-let canDownloadAppAssets: boolean | undefined;
-let canRVMLaunchAndDownload: boolean | undefined;
+
 /**
  * Initialize the snap provider.
  * @param options Options for the snap provider.
@@ -65,11 +64,10 @@ export async function init(options: SnapProviderOptions | undefined): Promise<vo
 			return;
 		}
 
-		await getCanDownloadAppAssets();
-		await getCanLaunchExternalProcess();
-		await isNativeSupportedByRVM();
+		const hasLaunchExternalProcess = await getCanLaunchExternalProcess(logger);
+		const hasDownloadAppAssets = await getCanDownloadAppAssets(logger);
 
-		if (!canDownloadAppAssets || !canLaunchExternalProcess || !canRVMLaunchAndDownload) {
+		if (!hasLaunchExternalProcess || !hasDownloadAppAssets) {
 			logger.warn(
 				"Snap is enabled but the platform does not have the capability or permission to download app assets or launch external processes."
 			);
@@ -335,63 +333,4 @@ export async function getAppAssetExecutablePath(appAssetInfo: {
 		localAppUrl = localAppUrl.slice(0, -1);
 	}
 	return [localAppUrl, "assets", appAssetInfo.alias, appAssetInfo.version, appAssetInfo.target].join(sep);
-}
-
-/**
- * Do we have the permissions to launch external processes.
- * @returns True if we have permission.
- */
-async function getCanLaunchExternalProcess(): Promise<boolean> {
-	if (!isEmpty(canLaunchExternalProcess)) {
-		return canLaunchExternalProcess;
-	}
-
-	try {
-		const canLaunchExternalProcessResponse = await fin.System.queryPermissionForCurrentContext(
-			"System.launchExternalProcess"
-		);
-
-		canLaunchExternalProcess = canLaunchExternalProcessResponse?.granted;
-	} catch (error) {
-		logger.error("Error while querying for System.launchExternalProcess permission", error);
-		canLaunchExternalProcess = false;
-	}
-
-	return canLaunchExternalProcess;
-}
-
-/**
- * Do we have the permissions to download app assets.
- * @returns True if we have permission.
- */
-async function getCanDownloadAppAssets(): Promise<boolean> {
-	if (!isEmpty(canDownloadAppAssets)) {
-		return canDownloadAppAssets;
-	}
-
-	try {
-		const canDownloadAppAssetsResponse =
-			await fin.System.queryPermissionForCurrentContext("System.downloadAsset");
-		canDownloadAppAssets = canDownloadAppAssetsResponse?.granted;
-	} catch (error) {
-		logger.error("Error while querying for System.downloadAsset permission", error);
-		canDownloadAppAssets = false;
-	}
-
-	return canDownloadAppAssets;
-}
-
-/**
- * The mac rvm as of version 11 does not support download app asset
- * or launch external process. Once there is support then we will
- * update logic to support a minimum rvm version on mac.
- * @returns True if the downloading and/or launching native apps is supported by the RVM.
- */
-async function isNativeSupportedByRVM(): Promise<boolean> {
-	if (!isEmpty(canRVMLaunchAndDownload)) {
-		return canRVMLaunchAndDownload;
-	}
-	const hostSpecs = await fin.System.getHostSpecs();
-	canRVMLaunchAndDownload = !hostSpecs.name.toLowerCase().startsWith("mac");
-	return canRVMLaunchAndDownload;
 }
